refactor(openai): 어댑터 패키지를 포트(기능) 단위로 재구성 — chat/title/summary/guardrail 분리, 공용 배관만 루트
