+++ src/main/java/com/readum/infrastructure/ai/openai/guardrail/PromptInjectionPatternAdvisor.java
@@ -1,4 +1,4 @@
-package com.readum.infrastructure.ai.openai.advisor;
+package com.readum.infrastructure.ai.openai.guardrail;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.ai.chat.client.ChatClientRequest;
